# internal/assets/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice Management API
  description: API for managing invoices, categories, companies, and file uploads
  version: 1.0.0
  contact:
    name: RxTech Lab
    url: https://github.com/rxtech-lab/invoice-management
servers:
  - url: /api
    description: API server
...
paths:
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: List categories
      description: Returns a paginated list of categories with optional search
      operationId: listCategories
      parameters:
        - name: keyword
          in: query
          description: Search keyword for category name
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Categories
      summary: Create category
      description: Creates a new invoice category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category
      description: Returns a category by ID
      operationId: getCategory
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Categories
      summary: Update category
      description: Updates an existing category
      operationId: updateCategory
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Categories
      summary: Delete category
      description: Deletes a category
      operationId: deleteCategory
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Category deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/companies:
    get:
      tags:
        - Companies
      summary: List companies
      description: Returns a paginated list of companies with optional search
      operationId: listCompanies
      parameters:
        - name: keyword
          in: query
          description: Search keyword for company name
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of companies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Companies
      summary: Create company
      description: Creates a new company/vendor
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/companies/{id}:
    get:
      tags:
        - Companies
      summary: Get company
      description: Returns a company by ID
      operationId: getCompany
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '200':
          description: Company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Companies
      summary: Update company
      description: Updates an existing company
      operationId: updateCompany
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Company updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Companies
      summary: Delete company
      description: Deletes a company
      operationId: deleteCompany
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '204':
          description: Company deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    CategoryId:
      name: id
      in: path
      required: true
      description: Category ID
      schema:
        type: integer
    CompanyId:
      name: id
      in: path
      required: true
      description: Company ID
      schema:
        type: integer
...
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
    Category:
      type: object
      properties:
        id:
          type: integer
          description: Category ID
        user_id:
          type: string
          description: Owner user ID
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
        color:
          type: string
          description: Hex color code (e.g., #FF5733)
          pattern: '^#[0-9A-Fa-f]{6}$'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
        color:
          type: string
          description: Hex color code
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
        color:
          type: string
          description: Hex color code
    CategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Company:
      type: object
      properties:
        id:
          type: integer
          description: Company ID
        user_id:
          type: string
          description: Owner user ID
        name:
          type: string
          description: Company name
        address:
          type: string
          description: Company address
        email:
          type: string
          format: email
          description: Contact email
        phone:
          type: string
          description: Contact phone
        website:
          type: string
          format: uri
          description: Company website
        tax_id:
          type: string
          description: Tax identification number
        notes:
          type: string
          description: Additional notes
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCompanyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Company name
        address:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
          format: uri
        tax_id:
          type: string
        notes:
          type: string
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
          format: uri
        tax_id:
          type: string
        notes:
          type: string
    CompanyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Company'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
security:
  - BearerAuth: []